requirements.txt
# Core dependencies
asyncio
aiohttp>=3.8.0
websockets>=10.0
pydantic>=2.0.0
pyyaml>=6.0

# Trading libraries
ccxt>=4.0.0
hyperliquid-python-sdk>=0.1.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.0
psutil>=5.9.0

# Logging and monitoring
structlog>=23.0.0
python-dotenv>=1.0.0

# Optional: Database support
sqlalchemy>=2.0.0
alembic>=1.12.0

# Optional: Notifications
discord.py>=2.3.0
slack-sdk>=3.21.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Monitoring
prometheus-client>=0.17.0